tests/wpt/tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_flex-grow-invalid.html

<!DOCTYPE html>
<title>flexbox | computed style | flex-grow: negative</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#flex-grow-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	flex-grow: -1;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-grow"),
	          "0");
});
</script>
